Change map coordinate variables in pf_map structure from 'short' to 'int'. This allows them to be passed to coordinate functions like normalize_map_pos.
Issued as PR#3731.

[[originally from CVS; cvs2svn created svn r6086]]

+++ common/aicore/path_finding.c
@@ -98,7 +98,7 @@ enum pf_node_status {
  * (aka internal buffer); index is map_pos_to_index(x, y);
  */
 struct pf_map {
-  short x, y;			/* The current position */
+  int x, y;			/* The current position */
   index_t index;		/* Current offset into lattice */
   struct pf_parameter *params;  /* Initial parameters */
   struct pqueue *queue;         /* Queue of nodes we have reached but not
